requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn>=21.0.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
werkzeug>=3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
bcrypt==4.1.2
google-auth==2.29.0

